chto-to

+++ srcs/parse/parse_cone.c
@@ -6,7 +6,7 @@
 /*   By: f0rsunka <f0rsunka@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/06/09 19:42:26 by f0rsunka          #+#    #+#             */
-/*   Updated: 2020/06/10 12:47:57 by f0rsunka         ###   ########.fr       */
+/*   Updated: 2020/06/10 17:06:17 by f0rsunka         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 t_cone		*create_cone(int fd, char **line)
 {
 	t_cone	*cone;
-	short	bitmask;
+	int		bitmask;
 	size_t	i;
 
 	bitmask = 0;
